requirements.txt
groq
gradio
python-dotenv
youtube-transcript-api
